+++ web/app/api/strava/[activityType]/route.ts
@@ -1,6 +1,6 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { activityToStravaTypes } from '../utils';
-import { STRAVA_API_BASE_URL } from '../configs';
+import { STRAVA_API_BASE_URL } from '../config';
 
 /**
  * @param req
